+++ extensions/quit-applications/src/errors.ts
@@ -1,0 +1,3 @@
+export class CancelledByUserError extends Error {
+  name: string = "CancelledByUserError";
+}