+++ extensions/quit-applications/src/utils/index.ts
@@ -1,0 +1,3 @@
+export * as exec from "./shell";
+export * as common from "./common";
+export * as widget from "./widget";